requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
crispy-bootstrap5==0.7
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.2
django-allauth==0.51.0
django-bootstrap5==22.1
django-crispy-forms==1.14.0
django-extensions==3.2.1
djangorestframework==3.14.0
env==0.1.0
idna==3.4
importlib-metadata==5.0.0
install==1.3.5
Markdown==3.4.1
netifaces==0.11.0
oauthlib==3.2.1
Pillow==9.2.0
pycparser==2.21
pydot==1.4.2
PyJWT==2.5.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
types-cryptography==3.3.23.1
urllib3==1.26.12
validate-email==1.3
zipp==3.9.0
